Remove duplication between partials showing moves and ideas
